# Only call ExportedFunction once