This is important ever since the call engine refactoring.

Did actually notice it in code review but wasn't confident without the benchmark to confirm :)

After
```
Benchmark_httpwasm_tinygo
Benchmark_httpwasm_tinygo-10    	  149740	      7940 ns/op
Benchmark_httpwasm_wat
Benchmark_httpwasm_wat-10       	  194221	      6163 ns/op
```

Before
```
Benchmark_httpwasm_tinygo
Benchmark_httpwasm_tinygo-10    	  125712	      9435 ns/op
Benchmark_httpwasm_wat
Benchmark_httpwasm_wat-10       	  136298	      7470 ns/op
```