@codefromthecrypt 2022-10-05
I think this is currently safe. I know @mathetake had recently done some refactoring and during that we were saying people shouldn't really cache functions, but in practice they still do.